t/t-unlock.sh: test unlock by path without refspec
As described in a previous commit in this PR, since PR #2773 the
lfstest-gitserver.go test server can mimic a Git LFS locking API
service which either does or does not read a "refspec" query parameter
and use it to filter results when searching for a lock matching a specific
file path.

Four tests of the "git lfs unlock" command were added or modified in
that PR which all require the test server to check for a "refspec" query
parameter when responding to a request to list matching locks.  These
tests signal the test server to perform the "refspec" argument check
by using test repository names that contain the special keyword suffix
"branch-required".  (However, these tests did not exercise that test
server behaviour because they all passed the --id option to the "git lfs
unlock" command, thus causing it to make no lock listing requests to
the test server.)

In the previous commit in this PR where we described these tests and
their history, we corrected them and added two more of the same kind
which also require the test server to filter search results using
the "refspec" query parameter.

However, no tests were added in PR #2773 which explicitly test the
"git lfs unlock" command against an API service that did not require
the "refspec" query parameter when searching for matching locks.  Some
tests, such as the "unlocking a lock (--json)" test, do so implicitly,
but none check all aspects of the expected behaviour in this case.

We therefore add a pair of tests, the first of which simply performs a
"git lfs unlock" command using a file path argument and confirming
it succeeds when the test server emulates an older API service that
does not check the "refspec" query argument in lock listing requests.
This test is akin to others, like the "unlocking a lock (--json)" test,
but explicitly named to make its intent clear.

The second new test we add checks that a "git lfs unlock" command which
uses a file path argument succeeds and removes a lock even if it passes
a "refspec" query parameter to the server that does not match the ref
name sent in the original lock request, so long as the test server
is emulating an older API service and does not recognize or check the
"refspec" argument in lock listing requests.  Note that this new
"unlocking a lock by path with bad ref without a ref required" test
is similar to the "unlocking a lock by path with bad ref" test, but
because it configures the test server to behave like an older API
service, it succeeds in removing the lock, whereas the "unlocking a
lock by path with bad ref" test expects to not remove the lock due to
the mismatched refs in its lock and unlock requests.