track: return an error on error or no action
If the user requests a pattern that matches `.gitattributes` and that's
the only pattern, we currently return a zero exit status, even though
we've failed.  Similarly, if an error occurs, we fail to exit nonzero.
In both of these cases, switch to exiting nonzero to indicate that the
operation was unsuccessful.

However, continue to return successfully if the reason we didn't process
any patterns was because they're already matched, since what the user
wanted is already done.

Co-authored-by: Chris Darroch <chrisd8088@github.com>