go.*,lfs: mock time in copy callback log file test
In commit 00f1e9521ab5d39aab393796745189d39c27b9a0 of PR #5504 we added
the TestCopyCallbackFileThrottle() test function to confirm that the
CopyCallbackFile() method of GitFilter structure in the "lfs" package
now throttled its output of progress messages into a log file so that,
in general, no more than one message was written during the interval
defined by the tasklog.DefaultLoggingThrottle value.  This new throttling
behaviour was added in commit 5bfa9009dafce59789b88b1a0854691b596f6e76
in the same PR.

The TestCopyCallbackFileThrottle() function attempts to validate
the throttling action by reading a test buffer in segments and waiting
or not waiting between each read, expecting the log messages to be
written or not written, as appropriate.  However, because our CI test
suite may experience slow execution times as a result of running in
virtualized environments, reads may occur at times more than the
tasklog.DefaultLoggingThrottle interval apart even without the artifical
waits injected by the test function, causing the test to fail when it
sees more log messages than it expects.

We can avoid this problem by making the wait intervals in the test
deterministic using a mocked value for the current time.  We use
the github.com/jmhodges/clock package for this purpose, adding an
object of its Clock interface type to the GitFilter structure.  In
all non-test code this object is simply a wrapper for the "time"
package's functions, but in our test we can use a variant which allows
us to control the time values returned to the CopyCallbackFile() method,
either stepping forward or not, as we desire.