ssh: Specifically designate a master multiplex connection
SSHTransfer.Shutdown() was attempting to first shut down the first
connection it created (which happened to be the master connection),
but this deadlocked because the master connection was waiting for
the extra connections to shut down. Designate the first connection
as the master connection, make sure that it truly is the master
connection, and shut it down after shutting down all extra
connections.

Issue: #5535