Add a script to notarize macOS binaries
We'll want to notarize macOS binaries using the new `notarytool` command
rather than `gon` because the latter uses `altool` under the hood, which
will be disabled as of November 1, 2023.  Create a script that runs the
proper command without echoing the arguments so that we don't include the
password in the log, even if it is filtered.