Add the original LFS URL to the Endpoint struct
When we compute the contents of an Endpoint struct, we don't preserve
the URL from which we computed the value in the first place.  For
example, if we use an SSH URL, we don't preserve the SSH URL anywhere
for further use.  Let's do so, since we'll use it in a future commit.

Note that for anything but an SSH URL, this is the same value as the URL
field.