locks: avoid looking up root directory repeatedly
If we're locking or unlocking multiple files, we don't need to look up
the top-level directory for the repository repeatedly.  We've almost
certainly done so when we've looked up the configuration, so let's save
that information in the fields we already have and use that to compute
the absolute path.