git: allow validating remote from a cached list
Right now, every time we call `ValidateRemote`, we look up the list of
remotes by invoking `git remote`.  It would be nice if we could cache
the list of remotes so we could validate the remote multiple times
without the expense of calling Git multiple times, so let's add a
function, `ValidateRemoteFromList`, which can take a cached list and
perform the same computation.