lfsapi: cache the remote list for each endpoint lookup
When we're locking or unlocking multiple times, we may end up looking up
the endpoint a few times.  This is very cheap if we don't need to invoke
Git.  However, as written right now, we call Git every time we validate
a remote, which is wasteful.

To make this more efficient, let's compute the remote list once and
cache it to avoid spawning a process.  Then, we can use the cached value
every time we need to validate a remote.