lfsapi: avoid invoking Git on every endpoint lookup
Right now, we call `GitDir` to find the Git directory for our repository
every time we look up an endpoint.  However, we _might_ already have
this data in our Git configuration handler, so let's use that if
possible so we don't need to spawn a process every time.  If it's not,
then cache the data so that we spawn at most one process.