Makefile: split release signing workflow for Windows
In the future, we'll want to use Azure Code Signing to sign our
binaries.  In order to do so successfully, we'll need to use a GitHub
Action to sign these binaries.  The Action we'll be using only allows
selecting binaries by directory and file extension; all binaries with a
given extension in a directory will be signed.

To make this easier, split the signing into several stages.  First, we
need to sign the three Git LFS binaries themselves, which is the first
stage, and then, once they're bundled into the installer, the second
stage will sign the installer itself.  Place these files in separate
directories to make signing them work with our Action.  Finally,
reassemble all the binaries into a tar archive for transport to our
Linux Action which will assemble them into a release, just as before.