script/changelog: add option for patch releases
When we prepare a patch release of Git LFS, we cherry-pick specific merged
changes from the "main" branch onto a "release-next" branch using the
"git cherry-pick" command with the -m1 option, which flattens all the
commits unique to the merge on the "main" branch into a single non-merge
commit on the release branch.

As a result, when we run our script/changelog script to generate a
summary of the changes for the patch release, the "git rev-list" command
it executes returns nothing because it uses the --merges option, and
the flattened commits on the release branch are not actual merge commits.

We therefore add a --patch option (as suggested by bk2204) to our script
which runs "git rev-list" without the --merges option.

We also add the -s option to the "git show" command the script runs
so as to avoid cluttering the output with the patch diff for each
commit it presents for categorization by the user.  When run for a
non-patch Git LFS release, this is not important because the commits
found by "git rev-list" are all merges and generally have no blob
changes directly associated with them, so the "git show" output is
relatively succinct.  But with the commits on the release branch for
a Git LFS patch release, which are not merges, the "git show" output
without the -s option is potentially quite verbose.