t/testhelpers: move credential setup into a function
Right now, we do the default credential setup once for each run of the
gitserver.  However, in the future, we'll want to be able to perform it
again in a different directory so we can have test-specific credential
directories if needed.  To do that, let's move the setup into a
function, which we can reuse elsewhere.