t/clone: fix cloneSSL test
This test has been broken for a while, but we hadn't noticed because of
an unexpected way of skipping the test for TravisCI, which we no longer
use.

In TravisCI, the `TRAVIS` environment variable is set to 1.  However,
in other environments, it is usually not set.  As a consequence, it
expands to the empty string.

The `if` statement in sh is not the `if` statement in C, and thus it
does not take a boolean value, and "1" would not be a valid value here.
The empty string is a valid value, though, and it is interpreted as a
no-op command, much like `:` or `true`, which are the customary ways to
write this.  Because this command always succeeds, we always skip this
branch, and the body of the test is not executed.

After removing this, our test fails because `assert_clean_status` is not
run in a Git repository.  We clearly wanted to run it in `$newclonedir`,
so let's do that.  In addition, the text "Git LFS:" is no longer
printed, so don't look for it.  With those changes, this test now
passes.