tools/util_windows.go: always close cloned files
In commit e861b79a49b02996d22191ebb142feb8768358de of #5595 the
Windows variant of our tools.CheckCloneFileSupported() function was
updated to address the problem that it could fail to remove the
temporary files it creates to test whether the "file clone" operation
(as used by the "git lfs dedup" command) is supported in a given
directory.  This problem could occur on Windows because we did not
close all of the temporary file's open descriptors, and so our
deferred call to os.Remove() might not succeed.

In a similar vein, the Windows variant of our tools.CloneFileByPath()
function opens both the source and destination file paths provided for
a "clone" operation but does not close the resultant file descriptors,
so we now add deferred calls to ensure we always close both files.
(This minor issue dates from the introduction in commit
e55bc4c5d41c227515348ef126743806d7e6a0c2 of the support for the
"file clone" operation on ReFS in PR #3790.)