tools/util_{darwin,linux}.go: close cloned files
In commit e861b79a49b02996d22191ebb142feb8768358de of #5595 the
Windows variant of our tools.CheckCloneFileSupported() function was
updated to address the problem that it could fail to remove the
temporary files it creates to test whether the "file clone" operation
(as used by the "git lfs dedup" command) is supported in a given
directory.  This problem could occur on Windows because we did not
close all of the temporary file's open descriptors, and so our
deferred call to os.Remove() might not succeed.

Although this issue does not affect non-Windows systems, as they
generally permit files to be deleted before all open descriptors
for the files are closed, we can keep our implementations aligned
and consistent if we also make sure to close any files we open
in the Linux and macOS variants of our tools.CheckCloneFileSupported()
and tools.CloneFileByPath() functions.

Note that on macOS, the CheckCloneFileSupported() function
creates two temporary files and then only passes them by name
to the CloneFileByPath() function, so we can close them immediately
after they are opened.  This is because the tools.CloneFile()
function is not supported on macOS.  By contrast, on Linux, the
CheckCloneFileSupported() function uses CloneFile(), so it must
defer closing the temporary files it creates rather than close
them directly after opening them.