tools/iotools.go: always close temp spool file
In commit e861b79a49b02996d22191ebb142feb8768358de of #5595 the
Windows variant of our tools.CheckCloneFileSupported() function was
updated to address the problem that it could fail to remove the
temporary files it creates to test whether the "file clone" operation
(as used by the "git lfs dedup" command) is supported in a given
directory.  This problem could occur on Windows because we did not
close all of the temporary file's open descriptors, and so our
deferred call to os.Remove() might not succeed.

In a similar vein, we also update our tools.Spool() function (which
is used in the "git lfs smudge" command to handle non-pointer data) to
ensure that if it creates a temporary file into which it spools data,
it always closes that file before trying to remove it.  While this
should have no advantage on non-Windows systems which allow for files
to be deleted while there are still open file descriptors for them,
on Windows this change should mean we never fail to clean up our
temporary spool file, if we create one.