t/t-smudge.sh: test smudge spooling to temp file
In a prior commit in this PR we updated the tools.Spool() function,
which is used only by the "git lfs smudge" command and only when it
must pass through non-pointer data.  When this data is longer than
the default internal memory buffer of 1024 bytes, the Spool() function
creates a temporary file to buffer the data, and then defers an
os.Remove() call to delete the file when the operation is complete.

At present this use of a temporary file when spooling is not tested by
the t/t-smudge.sh test suite, so we add a check in the existing "smudge
with invalid pointer" test which forces the use of such a file by
sending more than 1024 bytes of non-pointer data.