workflows/release: use Azure Code Signing
All new keys for Windows code signing have to be hardware backed as of
sometime in 2024.  To make this possible for us, let's set up Azure Code
Signing.

We call each of the separate stages, signing the intermediate objects in
between each stage.  The Action requires that we specify all of the
objects in a single folder with a single extension, so we separate the
binaries of Git LFS itself from the installer binary so we can include
signed binaries in the installer without re-signing the original
binaries.