tq: lazily initialize PktlineConnection
Right now, we spawn potentially several connections without necessarily
needing to.  For example, if we're transferring two objects, we can
practically use at most three connections: one for the batch and one for
each object.

Instead of spawning all the workers up front, let's change how we do
things and simply pass in the worker number at first.  Then, acquire the
connection when we're uploading or downloading so that the connection is
only spawned on demand.  This means that we spawn fewer workers in case
there are few objects.