attribute: warn if config exists after uninstalling
If someone has a system configuration set up for Git LFS, such as with
their package manager, `git lfs uninstall` will not be effective unless
the user has specified that file and has privileges to edit it.  In such
a case, let's warn the user so that they know that their attempt at
uninstalling might not have been effective.