t: add tests for invalid refs
The "git lfs fetch", "git lfs fsck", and "git lfs migrate" commands
detect invalid refs provided as command-line arguments and report them as
errors, but as we do not have tests for this behaviour, we add some now.

(Note that the "git lfs fetch" command also writes a stack trace to a
log file, so we could check that as well.  However, this action dates
from the early implementations of the precursor "git lfs get" command
and is now somewhat excessive compared to how we handle invalid ref
argument conditions in other commands, and so we may revise this
behaviour in the future.  Given that, the test we implement here should
be sufficient.)

Co-authored-by: Philip Peterson <philip.c.peterson@gmail.com>