Move SLES and OpenSUSE 15 to Rocky 8
The versions of SUSE and OpenSUSE 15 that are supported are actually
quite modern, with things like glibc 2.31.  It's safe to assume that
the Rocky 8 packages will function just fine on these versions and this
has been confirmed by installing the v3.4.1 Rocky 8 packages on SUSE
15.5.

Leave OpenSUSE 12 on CentOS 7, since it's a much older version.

Note that this update becomes relevant in the next release after v3.5,
when we'll almost certainly have hit end of life for CentOS 7 and will
want to remove it.