t/t-ls-files.sh: test files in subdirectory
The "git lfs ls-files" command was added in PR #122 (technically,
the precursor "git media ls-files" command) and its test suite
converted to shell tests in PR #336; however, the basic functionality
of the command has never had tests which confirm it handles files
in subdirectories appropriately.  (Some tests added in later PRs which
do check files in subdirectories under specific conditions, such as
with the --exclude option or with non-ASCII characters in subdirectory
names.)

As we expect to expand this command's test suite in subsequent commits
in this PR, we first add a new test which simply confirms that the
normal output of the command, and the output with the --debug option,
perform as expected when files have been both added and removed within
a subdirectory.

We also add a test which confirms the same behaviour when the --json
option is used.  The use of a separate test for this option was
preferred in PR #5007 when the --json option was first introduced (rather
than overloading the existing tests, as was done for the --debug option
when it was added in PR #2540), so we follow that model here as well.