t/t-ls-files.sh: test checkout and download status
In commit 2082b26ce728bb8a960b28bcac9aaecbb6a251c1 of PR #681 the
"git lfs ls-files" command was enhanced to include in its normal
output a simple symbolic indicator of whether a full copy of a Git
LFS object existed as a file in the working tree, with a "*" character
indicating the presence of such a file, and a "-" indicating the file's
absence.  While some tests were added at this time, none were added which
check the command's output when a file is absent from the working tree.

Later, in commit 1565e631e8914a619d13d28e1839847f506b428e of PR #2540,
the --debug option was added to the "git lfs ls-files" command, which
provides more extensive output, including separate Boolean fields
indicating whether a Git LFS object exists as either a file in the
working tree or as a cached object file.  Again, while some tests were
added for this new functionality, they do not check the cases where files
are absent from either the working tree or the Git LFS object cache.

As we expect to resolve a bug with the display of these data files
in a subsequent commit in this PR, we first introduce a new test
which validates the existing (and correct) behaviour.  (Notably,
this test only runs the "git lfs ls-files" command in the root of the
working tree; if we executed the command while in a subdirectory,
the output would not be correct, and this is the bug we will resolve
in another commit in this PR.)

We also add a test which confirms the same behaviour when the --json
option is used.  The use of a separate test for this option was
preferred in PR #5007 when the --json option was first introduced (rather
than overloading the existing tests, as was done for the --debug option
when it was added in PR #2540), so we follow that model here as well.