commands,t: fix output of ls-files in subdirectory
As reported in the following issue comment, the "git lfs ls-files"
command does not report the presence or absence of Git LFS object files
in the working tree correctly unless the command is run the root
directory of the working tree:

https://github.com/git-lfs/git-lfs/issues/1189#issuecomment-1950664657

We resolve this problem by ensuring that the full, canonicalized path
to the working tree is prepended by the fileExistsOfSize() helper
function to the paths of the files whose presence it checks with the
os.Stat() function.

We then add two new tests which validate the correct behaviour of the
command when it is run in a subdirectory of the working tree.  These
tests would fail without the accompanying bug fix to the command.

We also add another two new tests which confirm the same behaviour when
the --json option is used.  The use of a separate test for this option was
preferred in PR #5007 when the --json option was first introduced (rather
than overloading the existing tests, as was done for the --debug option
when it was added in PR #2540), so we follow that model here as well.