lfshttp: return errors for certificate problems
Right now, our certificate handling silently ignores any problems and
just pretends like no certificate was specified.  We do log these values
to the trace output, but it's hard for people to find out what's going
on and correct the problem.

Let's add some error handling here so that people know what's going
wrong and can act on it.  In this particular case, we include specific
information in the key file that indicates that we only accept PKCS#1
files to help the user debug.