lfshttp: expand SSL key and cert paths
Git accepts paths for SSL keys and certs starting with a tilde to expand
to the user's home directory.  This is very useful for configuration
which must be shared across systems.  Let's add support for the same
thing here and add a test to make sure it works.

Skip the test on Windows, because Git for Windows returns “fatal:
refusing to work with credential missing host field” when it sees the
fake `cert:///` URL without a hostname that's used to query the
credential help for our TLS certificate credentials.