.github/workflows: install msgfmt on Windows
As of Git for Windows v2.44, the msgfmt.exe binary is no longer
installed by default and is therefore not available in the GitHub
Actions runners we use for our CI and release jobs.  We instead
use the git-for-windows/setup-git-for-windows-sdk Action to install
the SDK for Git for Windows, which includes msgfmt.exe.

h/t dscho for the advice to use the SDK!