Build release assets with Go 1.21
Due to the issue noted in https://github.com/golang/go/issues/65705, we
don't want to build our release assets with 1.22.0, since this can cause
deadlock in the HTTP client.  Instead, let's keep testing with both 1.21
and 1.22 in our CI jobs, but build with 1.21 for release.

In the future, once that bug is fixed and an updated version is
released, we can move back to 1.22.