Include remote error when pure SSH protocol fails
Right now, if spawning the remote command fails for some reason, we get
a message like the following in the trace output:

  pure SSH protocol connection failed: Unable to negotiate version with remote side (unable to read capabilities): EOF

However, all we know is that we failed to connect to the remote
`git-lfs-transfer` process.  That could be for any number of reasons:
the remote program doesn't exist or is not in `PATH`, the user has
misconfigured something locally or remotely, or any number of other
problems.  We won't know unless we log the output, so let's do that so
we can get more information.