Use a git-update-ref script to update references
Invoking git-update-ref to rewrite a single reference is a noticeable
performance bottleneck during LFS migration in a repository with a lot
of references. Particularly, it may take ~3 minutes to update ~20k
references.

The patch makes Git LFS update all references at once by generating
a reference update script and piping it into git-update-ref.

Please note that logic that handles nested tags in updateOneRef()
assumes that the tag referenced by the tag being updated has already
been updated in the Git storage, this assumption no longer holds.
The patch tackles the issue by caching the new state of the reference
in the map 'seen'.